requirements.txt
mcp>=0.1.0
splunk-sdk>=2.0.0
cryptography>=41.0.0
pyyaml>=6.0
requests>=2.31.0
urllib3>=2.0.0
